move errors into separate file
